Implement support to removeSkills
Refactor ProfileService

Pending: adjust client-side to support new View Models.
