--- app.web/Models/Page/Menu.cs
@@ -1,12 +1,0 @@
-namespace app.web.Models
-{
-    using System.Collections.Generic;
-    using System.Linq;
-
-    public class Menu
-    {
-        public MenuItem LeftMenu { get; set; }
-
-        public MenuItem RightMenu { get; set; }
-    }
-}
