--- app.web/Models/Page/Page.cs
@@ -1,7 +1,0 @@
-namespace app.web.Models
-{
-    public class Page
-    {
-        public Menu Menu { get; set; }
-    }
-}
