--- app.web/Services/Page/IPageConfigurationService.cs
@@ -1,9 +1,0 @@
-namespace app.web.Services
-{
-    using System.Threading.Tasks;
-
-    public interface IPageConfigurationService
-    {
-        Task<PageConfiguration> GetPageConfiguration();
-    }
-}
