--- app.web/Services/Page/IPageService.cs
@@ -1,10 +1,0 @@
-namespace app.web.Services
-{
-    using System.Threading.Tasks;
-    using app.web.Models;
-
-    public interface IPageService
-    {
-        Task<Page> GetPage();
-    }
-}
