--- app.web/Services/Page/Models/PageSide.cs
@@ -1,8 +1,0 @@
-namespace app.web.Services
-{
-    public enum PageSide
-    {
-        Left = 0,
-        Right = 1
-    }
-}
